backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
ldap3==2.9.1
cryptography==43.0.1
itsdangerous==2.1.2
